Backend/requirements.txt
alembic==1.7.6
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
boto3==1.20.27
botocore==1.23.27
cffi==1.15.0
click==8.0.3
colorama==0.4.4
contextlib2==21.6.0
contextvars==2.4
cryptography==36.0.1
dataclasses==0.6
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.3
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
importlib-resources==5.4.0
jmespath==0.10.0
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
Pillow==8.4.0
psycopg2==2.9.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
python-http-client==3.3.5
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
s3transfer==0.5.0
sendgrid==6.9.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starkbank-ecdsa==2.0.3
starlette==0.16.0
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
watchgod==0.7
websockets==9.1
Werkzeug==2.0.2
zipp==3.6.0
